refactored duplicated code to load service configs
