+++ paasta_tools/cli/cmds/local_run.py
@@ -27,10 +27,10 @@ import service_configuration_lib
 from docker import Client
 from docker import errors
 
-from paasta_tools.chronos_tools import load_chronos_job_config
 from paasta_tools.cli.cmds.check import makefile_responds_to
 from paasta_tools.cli.cmds.cook_image import paasta_cook_image
 from paasta_tools.cli.utils import figure_out_service_name
+from paasta_tools.cli.utils import get_instance_config
 from paasta_tools.cli.utils import guess_cluster
 from paasta_tools.cli.utils import guess_instance
 from paasta_tools.cli.utils import lazy_choices_completer
@@ -38,7 +38,6 @@ from paasta_tools.cli.utils import list_instances
 from paasta_tools.cli.utils import list_services
 from paasta_tools.marathon_tools import CONTAINER_PORT
 from paasta_tools.marathon_tools import get_healthcheck_for_instance
-from paasta_tools.marathon_tools import load_marathon_service_config
 from paasta_tools.paasta_execute_docker_command import execute_in_container
 from paasta_tools.utils import _run
 from paasta_tools.utils import get_docker_host
@@ -51,7 +50,6 @@ from paasta_tools.utils import PaastaNotConfiguredError
 from paasta_tools.utils import SystemPaastaConfig
 from paasta_tools.utils import Timeout
 from paasta_tools.utils import TimeoutError
-from paasta_tools.utils import validate_service_instance
 
 BAD_PORT_WARNING = 'This_service_is_listening_on_the_PORT_variable__You_must_use_8888__see_y/paasta_deploy'
 
@@ -545,33 +543,6 @@ def run_docker_container(
         returncode = docker_client.inspect_container(container_id)['State']['ExitCode']
         _cleanup_container(docker_client, container_id)
     sys.exit(returncode)
-
-
-def get_instance_config(service, instance, cluster, soa_dir, load_deployments=False):
-    """ Returns the InstanceConfig object for whatever type of instance
-    it is. (chronos or marathon) """
-    instance_type = validate_service_instance(
-        service=service,
-        instance=instance,
-        cluster=cluster,
-        soa_dir=soa_dir,
-    )
-    if instance_type == 'marathon':
-        instance_config_load_function = load_marathon_service_config
-    elif instance_type == 'chronos':
-        instance_config_load_function = load_chronos_job_config
-    else:
-        raise NotImplementedError(
-            "instance is %s of type %s which is not supported by local-run"
-            % (instance, instance_type)
-        )
-    return instance_config_load_function(
-        service=service,
-        instance=instance,
-        cluster=cluster,
-        load_deployments=load_deployments,
-        soa_dir=soa_dir
-    )
 
 
 def configure_and_run_docker_container(docker_client, docker_hash, service, instance, cluster, args, pull_image=False):
